refactor: streamline oneWay functionality and remove legacy code; add runtime sidecar management
