--- src/oneWay.js
@@ -1,5 +1,0 @@
-export {
-	ONE_WAY_MODES,
-	oneWayAllowsDiscordToWhatsApp,
-	oneWayAllowsWhatsAppToDiscord,
-} from "./contracts.js";
